% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-containers.R
\name{build_micromamba_image}
\alias{build_micromamba_image}
\title{Build Micromamba Container Image}
\usage{
build_micromamba_image(
  dockerfile_path = NULL,
  image_name = "luciorq/condathis-micromamba:latest",
  force = FALSE,
  method = "docker"
)
}
\arguments{
\item{dockerfile_path}{Character. Path to Dockerfile.}

\item{image_name}{Character. Image name used for the contanier image.
Defaults to \code{"condathis-micromamba:latest"}.}

\item{force}{Logical. Should image be removed before building.
Defaults to FALSE.}

\item{method}{Character. One of \code{c("docker", "singularity")}.}
}
\description{
Build Micromamba Container Image
}
