% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-containers.R
\name{build_micromamba_image_singularity}
\alias{build_micromamba_image_singularity}
\title{Build Singularity / Apptainer Image}
\usage{
build_micromamba_image_singularity(
  image_name = "luciorq/condathis-micromamba:latest",
  registry_name = "docker"
)
}
\arguments{
\item{image_name}{Character. Image name used for the container image,
containing the remote repository, separated by forward slash.
Defaults to \code{"luciorq/condathis-micromamba:latest"}.}

\item{registry_name}{Character. Container Registry where image already
exists. Defaults to \code{"docker"} (Docker Hub).}
}
\description{
Build Singularity / Apptainer Image
}
