% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condathis-package.R
\docType{package}
\name{condathis-package}
\alias{condathis}
\alias{condathis-package}
\title{condathis: Condathis: run any CLI tool on a Conda Environment}
\description{
Condathis is an R package that simplifies the execution of command line tools within isolated conda environments. It enables users to effortlessly manage Conda environments, execute command line tools, handle dependencies, and ensure reproducibility in their data analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/condathis}
  \item Report bugs at \url{https://github.com/luciorq/condathis/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})

}
\keyword{internal}
