% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env_internal_docker}
\alias{create_env_internal_docker}
\title{Create Environment Using Docker}
\usage{
create_env_internal_docker(
  packages = NULL,
  env_file = NULL,
  env_name = "condathis-env",
  channels = c("bioconda", "conda-forge"),
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest",
  additional_channels = NULL,
  verbose = TRUE
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{env_file}{Character. Path to the YAML file with Conda Environment
description. If this argument is used, the \code{packages} argument should not
be included in the command.}

\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge")' are used for solving
dependencies.}

\item{container_name}{Character. Name of the Container created by Docker.
Defaults to \code{"condathis-micromamba-base"}.}

\item{image_name}{Character. Name of the Docker Image used, it will try
to pull it automatically if internet connection is available.
Defaults to \code{"luciorq/condathis-micromamba:latest"}.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}

\item{verbose}{Logical. Should command and messages be print to
the console.
defaults to TRUE.}
}
\description{
Create Environment Using Docker
}
