% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{define_method_to_use}
\alias{define_method_to_use}
\title{Automatically Find Suitable Infrastrucure to Run CLI Tools}
\usage{
define_method_to_use(
  packages = NULL,
  channels = NULL,
  additional_channels = NULL
)
}
\description{
Automatically Find Suitable Infrastrucure to Run CLI Tools
}
