% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_platform.R
\name{define_platform}
\alias{define_platform}
\title{Define Platform to be used by \code{create_env()}.}
\usage{
define_platform(
  packages,
  platform = NULL,
  channels = c("bioconda", "conda-forge", "defaults"),
  additional_channels = NULL
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{platform}{Character. Platform to search for \code{packages}.
Defaults to \code{NULL} which will use the current platform.
E.g. "linux-64", "linux-32", "osx-64", "win-64", "win-32", "noarch".
Note: on Apple Silicon MacOS will use "osx-64" instead of "osx-arm64"
if Rosetta 2 is available.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge", "defaults")' are used for solving
dependencies.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}
}
\description{
Define Platform to be used by \code{create_env()}.
}
