% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sys_arch.R
\name{get_sys_arch}
\alias{get_sys_arch}
\title{Get operating system and CPU architecture}
\usage{
get_sys_arch()
}
\value{
A character string such as \code{"Darwin-x86_64"} or
\code{"Linux-aarch64"}.
}
\description{
Returns the current operating system and CPU architecture as a single
string in the format \code{"<OS>-<Architecture>"}.
}
\examples{
# Retrieve the system architecture
condathis::get_sys_arch()
#> [1] "Darwin-x86_64"

}
