% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cmd_available.R
\name{is_cmd_available}
\alias{is_cmd_available}
\title{Check If Command Line Tool Is Available}
\usage{
is_cmd_available(cmd)
}
\arguments{
\item{cmd}{Character. The main command to be executed in the Conda environment.}
}
\description{
Test the required methods to check if command line tool is available on PATH,
Before trying to execute it.
}
