% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-containers.R
\name{is_singularity_available}
\alias{is_singularity_available}
\title{Are Singularity or Apptainer CLIs available}
\usage{
is_singularity_available()
}
\description{
Test if Singularity or Apptainer CLIs are available on PATH.
}
