% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_envs.R
\name{list_envs}
\alias{list_envs}
\title{List Conda environments managed by condathis}
\usage{
list_envs(verbose = "silent")
}
\arguments{
\item{verbose}{Character string controlling console output.
Defaults to \code{"silent"}.}
}
\value{
A character vector of environment names.
}
\description{
Returns environment names located under the \code{condathis} installation root.
Environments not managed by \code{condathis} are excluded.
}
\examples{
\dontrun{
condathis::with_sandbox_dir({
  # Create environments
  condathis::create_env(
    packages = "bioconda::fastqc",
    env_name = "fastqc-env"
  )
  condathis::create_env(
    packages = "python",
    env_name = "python-env"
  )

  # List environments
  condathis::list_envs()
  #> [1] "fastqc-env" "python-env"
})
}

}
