% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromamba_user_installed.R
\name{micromamba_user_installed}
\alias{micromamba_user_installed}
\title{Retrieve Micromamba Installed Path}
\usage{
micromamba_user_installed()
}
\description{
Checks if user already have \code{micromamba} binaries available on \code{PATH}.
}
\keyword{internal}
