% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_cmd.R
\name{native_cmd}
\alias{native_cmd}
\title{Run Micromamba Command}
\usage{
native_cmd(conda_cmd, conda_args = NULL, ..., verbose = FALSE, stdout = "|")
}
\arguments{
\item{conda_cmd}{Character. Conda subcommand to be run.
E.g. "create", "install", "env", "--help", "--version".}

\item{conda_args}{Character vector. Additional arguments passed to
the Conda command.}

\item{...}{Additional arguments to be used in the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special character or spaces.}

\item{verbose}{Logical. Should command and messages be print to
the console.
Defaults to \code{TRUE}.}

\item{stdout}{Default: "|" keep stdout to the R object
returned by \code{run()}.
A character string can be used to define a file path to be used as standard output. e.g: "output.txt".}
}
\description{
Run a command using micromamba executable in the native backend.
}
