% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{packages_search_docker}
\alias{packages_search_docker}
\title{Search for Packages in Channels using Docker}
\usage{
packages_search_docker(
  packages,
  channels = c("bioconda", "conda-forge", "defaults"),
  additional_channels = NULL,
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest"
)
}
\description{
Search for Packages in Channels using Docker
}
