% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages_search_native.R
\name{packages_search_native}
\alias{packages_search_native}
\title{Search For Package in Channels}
\usage{
packages_search_native(
  packages,
  channels = c("bioconda", "conda-forge"),
  platform = NULL,
  additional_channels = NULL,
  verbose = "silent"
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge")' are used for solving
dependencies.}

\item{platform}{Character. Platform to search for \code{packages}.
Defaults to \code{NULL} which will use the current platform.
E.g. "linux-64", "linux-32", "osx-64", "win-64", "win-32", "noarch".
Note: on Apple Silicon MacOS will use "osx-64" instead of "osx-arm64"
if Rosetta 2 is available and any of the \code{packages} is not available
for "osx-arm64".}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}

\item{verbose}{Character string specifying the verbosity level of the function's output. Acceptable values are:
\itemize{
\item \strong{"silent"}: Suppress all output from internal command-line tools. Equivalent to \code{FALSE}.
\item \strong{"cmd"}: Print the internal command(s) passed to the command-line tool.
\item \strong{"output"}: Print the standard output and error from the command-line tool to the screen. Note that the order of the standard output and error lines may not be correct, as standard output is typically buffered. If the standard output and/or error is redirected to a file or they are ignored, they will not be echoed.
\item \strong{"full"}: Print both the internal command(s) (\code{"cmd"}) and their standard output and error (\code{"output"}). Equivalent to \code{TRUE}.
Logical values \code{FALSE} and \code{TRUE} are also accepted for backward compatibility but are \emph{soft-deprecated}. Please use \code{"silent"} and \code{"full"} respectively instead.
}}
}
\description{
Search if Package is available in required Channels
}
\keyword{internal}
