% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_internal_docker}
\alias{run_internal_docker}
\title{Run Command Using Docker}
\usage{
run_internal_docker(
  cmd,
  ...,
  env_name = "condathis-env",
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest",
  mount_paths = NULL,
  gpu_container = FALSE,
  verbose = TRUE,
  stdout = "|"
)
}
\arguments{
\item{cmd}{Character. The main command to be executed in the Conda environment.}

\item{...}{Additional arguments to be passed to the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special characters or spaces.}

\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to \code{"condathis-env"}.
If the specified environment does not exist, it will be created automatically using \code{create_env()}.}

\item{container_name}{Character string. The name of the container to be used when running the command via Docker or Singularity. Defaults to \code{"condathis-micromamba-base"}.}

\item{image_name}{Character string. The name of the Docker image to use when the method is \code{"docker"}. Defaults to \code{"luciorq/condathis-micromamba:latest"}.}

\item{mount_paths}{Character vector. Host paths to be mounted in the container. Useful when using Docker or Singularity methods.}

\item{gpu_container}{Logical. Whether to enable GPU support in the container. Defaults to \code{FALSE}.}

\item{verbose}{Character string specifying the verbosity level of the function's output. Acceptable values are:
\itemize{
\item \strong{"silent"}: Suppress all output from internal command-line tools. Equivalent to \code{FALSE}.
\item \strong{"cmd"}: Print the internal command(s) passed to the command-line tool.
\item \strong{"output"}: Print the standard output and error from the command-line tool to the screen. Note that the order of the standard output and error lines may not be correct, as standard output is typically buffered. If the standard output and/or error is redirected to a file or they are ignored, they will not be echoed.
\item \strong{"full"}: Print both the internal command(s) (\code{"cmd"}) and their standard output and error (\code{"output"}). Equivalent to \code{TRUE}.
}

Logical values \code{FALSE} and \code{TRUE} are also accepted for backward compatibility but are \emph{soft-deprecated}. Please use \code{"silent"} and \code{"full"} respectively instead.}

\item{stdout}{Default: "|" keep stdout to the R object
returned by \code{run()}.
A character string can be used to define a file path to be used as standard output. e.g: "output.txt".}
}
\description{
Internal function to run a command in a Conda environment using Docker.
}
