% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_internal_native}
\alias{run_internal_native}
\title{Run Command Using Native Method}
\usage{
run_internal_native(
  cmd,
  ...,
  env_name = "condathis-env",
  verbose = FALSE,
  error = c("cancel", "continue"),
  stdout = "|",
  stderr = "|"
)
}
\arguments{
\item{cmd}{Character. The main command to be executed in the Conda environment.}

\item{...}{Additional arguments to be passed to the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special characters or spaces.}

\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to \code{"condathis-env"}.
If the specified environment does not exist, it will be created automatically using \code{create_env()}.}

\item{verbose}{Logical. Whether to display detailed output messages. Defaults to \code{FALSE}.}

\item{error}{Character string. How to handle errors. Options are \code{"cancel"} or \code{"continue"}. Defaults to \code{"cancel"}.}

\item{stdout}{Default: "|" keep stdout to the R object
returned by \code{run()}.
A character string can be used to define a file path to be used as standard output. e.g: "output.txt".}
}
\description{
Internal function to run a command in a Conda environment using the native method.
}
