% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_internal_singularity}
\alias{run_internal_singularity}
\title{Run Command Using Singularity}
\usage{
run_internal_singularity(
  cmd,
  ...,
  env_name = "condathis-env",
  sif_image_path = NULL,
  mount_paths = NULL,
  gpu_container = FALSE,
  verbose = FALSE,
  stdout = "|"
)
}
\arguments{
\item{cmd}{Character. The main command to be executed in the Conda environment.}

\item{...}{Additional arguments to be passed to the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special characters or spaces.}

\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to \code{"condathis-env"}.
If the specified environment does not exist, it will be created automatically using \code{create_env()}.}

\item{sif_image_path}{Character string. The file path to the Singularity image file (\code{.sif}) to use when the method is \code{"singularity"}.}

\item{mount_paths}{Character vector. Host paths to be mounted in the container. Useful when using Docker or Singularity methods.}

\item{gpu_container}{Logical. Whether to enable GPU support in the container. Defaults to \code{FALSE}.}

\item{verbose}{Logical. Whether to display detailed output messages. Defaults to \code{FALSE}.}

\item{stdout}{Default: "|" keep stderr to the R object
returned by \code{run()}.
A character string can be used to define a file path to be used as standard error. e.g: "error.txt".}
}
\description{
Internal function to run a command in a Conda environment using Singularity.
}
