% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromamba_user_installed.R
\name{symlink_micromamba_bin}
\alias{symlink_micromamba_bin}
\title{Retrieve Micromamba Installed Path}
\usage{
symlink_micromamba_bin(path = NULL, force = FALSE)
}
\description{
Checks if user already have \code{micromamba} binaries available on \code{PATH}.
}
\keyword{internal}
