% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_sandbox_dir.R
\name{with_sandbox_dir}
\alias{with_sandbox_dir}
\title{Execute code in an isolated temporary environment}
\usage{
with_sandbox_dir(code, .local_envir = base::parent.frame())
}
\arguments{
\item{code}{Expression to evaluate in the sandboxed environment.}

\item{.local_envir}{Environment used for local scoping and evaluation.
Defaults to \code{parent.frame()}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Evaluates code with temporary home, data, and cache directories.
This is mainly intended for examples and tests so no user files are written.
}
\examples{
condathis::with_sandbox_dir(print(fs::path_home()))
condathis::with_sandbox_dir(print(tools::R_user_dir("condathis")))
}
