% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condathis-package.R
\docType{package}
\name{condathis-package}
\alias{condathis}
\alias{condathis-package}
\title{condathis: Run Any CLI Tool on a 'Conda' Environment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simplifies the execution of command line interface (CLI) tools within isolated and reproducible environments. It enables users to effortlessly manage 'Conda' environments, execute command line tools, handle dependencies, and ensure reproducibility in their data analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/condathis}
  \item Report bugs at \url{https://github.com/luciorq/condathis/issues}
}

}
\author{
\strong{Maintainer}: Lucio Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Claudio Zanettini \email{claudio.zanettini@gmail.com} (\href{https://orcid.org/0000-0001-5043-8033}{ORCID}) [contributor]
}

}
\keyword{internal}
