% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_env}
\alias{create_env}
\title{Create Conda Environment with specific packages}
\usage{
create_env(packages, env_name)
}
\arguments{
\item{packages}{Character vector with the names of the packages and
version strings if necessary.}

\item{env_name}{Name of the Conda environment where the packages are
going to be installed.}
}
\description{
Create Conda Environment with specific packages
}
