% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create a Conda Environment}
\usage{
create_env(
  packages = NULL,
  env_file = NULL,
  env_name = "condathis-env",
  channels = c("bioconda", "conda-forge", "defaults"),
  method = c("native", "auto", "docker", "singularity"),
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest",
  sif_image_path = NULL,
  additional_channels = NULL,
  gpu_container = FALSE,
  platform = NULL,
  verbose = TRUE
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{env_file}{Character. Path to the YAML file with Conda Environment
description. If this argument is used, the \code{packages} argument should not
be included in the command.}

\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge", "defaults")' are used for solving
dependencies.}

\item{method}{Character. Beckend method to run \code{micromamba}, the default is
"auto" running "native" with the \code{micromamba} binaries installed
by \code{condathis}. Additional methods are supported for using Docker Linux Containers
"docker" and Singularity Containers "singularity" as the beckends,
those are especially useful if running on systems where the Conda
recipes are not available for the OS or CPU architecture in place.
The container-based backends leverage \code{dockerthis} R package.}

\item{container_name}{Character. Name of the Container created by Docker.
Defaults to \code{"condathis-micromamba-base"}.}

\item{image_name}{Character. Name of the Docker Image used, it will try
to pull it automatically if internet connection is available.
Defaults to \code{"luciorq/condathis-micromamba:latest"}.}

\item{sif_image_path}{Character. Path to SIF image file.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}

\item{gpu_container}{Logical. GPU support for Container Beckend \code{methods}.
This argument is not necessary if running native.
Default to FALSE.}

\item{platform}{Character. Platform to search for \code{packages}.
Defaults to \code{NULL} which will use the current platform.
E.g. "linux-64", "linux-32", "osx-64", "win-64", "win-32", "noarch".
Note: on Apple Silicon MacOS will use "osx-64" instead of "osx-arm64"
if Rosetta 2 is available.}

\item{verbose}{Logical. Should command and messages be print to
the console.
defaults to TRUE.}
}
\description{
Create Conda Environment with specific packages installed to be used by \code{run()}.
}
