% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env.R
\name{create_env}
\alias{create_env}
\title{Create Conda Environment with specific packages}
\usage{
create_env(packages = NULL, env_name = "condathis-env")
}
\arguments{
\item{packages}{Character vector with the names of the packages and
version strings if necessary.}

\item{env_name}{Name of the Conda environment where the packages are
going to be installed. Defaults to 'condathis-env'.}
}
\description{
Create Conda Environment with specific packages
}
