% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{define_method_to_use}
\alias{define_method_to_use}
\title{Automatically Find Suitable Infrastrucure to Run CLI Tools}
\usage{
define_method_to_use(
  packages = NULL,
  channels = c("bioconda", "conda-forge", "defaults"),
  additional_channels = NULL,
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest",
  sif_image_path = NULL
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge", "defaults")' are used for solving
dependencies.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}
}
\description{
Automatically Find Suitable Infrastrucure to Run CLI Tools
}
