% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_exists.R
\name{env_exists}
\alias{env_exists}
\title{Check If Environment Names Already exists}
\usage{
env_exists(env_name = "condathis-env")
}
\arguments{
\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}
}
\description{
Check If Environment Names Already exists
}
