% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_exists.R
\name{env_exists}
\alias{env_exists}
\title{Check If Environment Already exists}
\usage{
env_exists(env_name)
}
\arguments{
\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}
}
\value{
Boolean.
}
\description{
Check If Environment Already exists
}
\examples{
\dontrun{
# Create the environment
condathis::create_env(
  packages = "fastqc",
  env_name = "fastqc-env"
)
# Check if exists
condathis::env_exists("fastqc-env")
#> [1] TRUE
}
}
