% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_dir.R
\name{get_env_dir}
\alias{get_env_dir}
\title{Get an environment directory path}
\usage{
get_env_dir(env_name = "condathis-env")
}
\arguments{
\item{env_name}{Character string with the environment name.
Defaults to \code{"condathis-env"}.}
}
\value{
A character string with the expected environment directory path.
}
\description{
Returns the absolute path where an environment is expected under the
\code{condathis} installation root. The path is returned even if the
environment has not been created yet.
}
\examples{
condathis::with_sandbox_dir({
  # Get the default environment directory
  condathis::get_env_dir()
  #> "/path/to/condathis/envs/condathis-env"

  # Get the directory for a specific environment
  condathis::get_env_dir("my-env")
  #> "/path/to/condathis/envs/my-env"
})

}
