% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_env_dir.R
\name{get_env_dir}
\alias{get_env_dir}
\title{Retrieve Path To Environment}
\usage{
get_env_dir(env_name = "condathis-env")
}
\arguments{
\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}
}
\description{
Retrieve path to where environment should be created.
\strong{Note}: It retrieves the Path even if the environment is \strong{not} created yet.
}
