% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sys_arch.R
\name{get_sys_arch}
\alias{get_sys_arch}
\title{Return OS and CPU Architecture}
\usage{
get_sys_arch()
}
\value{
A character vector with one element indicating OS and CPU Architecture.
}
\description{
Return OS and CPU Architecture
}
\examples{
\dontrun{
# Create a Conda environment with the CLI fastqc
condathis::get_sys_arch()
#> [1] "Darwin-x86_64"
}
}
