% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_micromamba.R
\name{install_micromamba}
\alias{install_micromamba}
\title{Install Micromamba binaries in a \code{condathis} controlled path.}
\usage{
install_micromamba(timeout_limit = 3600, download_method = "auto")
}
\arguments{
\item{timeout_limit}{Timeout limit for downloading.}

\item{download_method}{Argument passed to \code{method} argument from
\code{utils::download.file()} function.}
}
\description{
Install Micromamba binaries in a \code{condathis} controlled path.
}
