% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_micromamba.R
\name{install_micromamba}
\alias{install_micromamba}
\title{Install Micromamba binaries in a \code{condathis} controlled path.}
\usage{
install_micromamba(
  timeout_limit = 3600,
  download_method = "auto",
  force = FALSE
)
}
\arguments{
\item{timeout_limit}{Numeric. Timeout limit for downloading the Micromamba
binaries, in seconds. Defaults to 3600 seconds (1 hour).}

\item{download_method}{Character. Argument passed to the \code{method} argument
of the \code{utils::download.file()} function used for downloading the binaries.
Defaults to "auto".}

\item{force}{Logical. If set to TRUE, the download and installation of the
Micromamba binaries will be forced, even if they already exist in the
system or \code{condathis} controlled path. Defaults to FALSE.}
}
\description{
This function downloads and installs the Micromamba binaries in the path
managed by the \code{condathis} package. Micromamba is a lightweight
implementation of the Conda package manager and provides an efficient way
to create and manage conda environments.
}
