% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install Packages in a Existing Conda Environment}
\usage{
install_packages(packages, env_name = "condathis-env")
}
\arguments{
\item{packages}{Character vector with the names of the packages and
version strings if necessary.}

\item{env_name}{Name of the Conda environment where the packages are
going to be installed. Defaults to 'condathis-env'.}
}
\description{
Install Packages in a Existing Conda Environment
}
