% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_envs}
\alias{list_envs}
\title{List Installed Environments}
\usage{
list_envs(verbose = FALSE)
}
\arguments{
\item{verbose}{Logical. Should command and messages be print to
the console.
defaults to TRUE.}
}
\description{
List Installed Environments
}
