% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List packages in a Conda environment}
\usage{
list_packages(
  env_name = "condathis-env",
  verbose = c("output", "silent", "cmd", "spinner", "full")
)
}
\arguments{
\item{env_name}{Character string with the target environment name.
Defaults to \code{"condathis-env"}.}

\item{verbose}{Character string controlling console output.
Supported values are \code{"output"}, \code{"silent"}, \code{"cmd"}, \code{"spinner"},
and \code{"full"}. Defaults to \code{"output"}.}
}
\value{
A data frame (\code{tibble}) with installed packages and the columns:
\itemize{
\item \strong{base_url}: The base URL of the package source.
\item \strong{build_number}: The build number of the package.
\item \strong{build_string}: The build string describing the package build details.
\item \strong{channel}: The channel from which the package was installed.
\item \strong{dist_name}: The distribution name of the package.
\item \strong{name}: The name of the package.
\item \strong{platform}: The platform for which the package is built.
\item \strong{version}: The version of the package.
}
}
\description{
Returns package metadata for a Conda environment as a tibble.
}
\examples{
\dontrun{
condathis::with_sandbox_dir({
  # Creates a Conda environment with the CLI `fastqc`
  condathis::create_env(
    packages = "bioconda::fastqc",
    env_name = "fastqc-env"
  )
  # Lists the packages in env `fastqc-env`
  dat <- condathis::list_packages("fastqc-env")
  dim(dat)
  #> [1] 34  8
})
}

}
