% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_packages}
\alias{list_packages}
\title{List Packages Installed Inside Environment}
\usage{
list_packages(env_name = "condathis-env")
}
\arguments{
\item{env_name}{Name of the Conda Environment where the tool is run.
Defaults to 'condathis-env'.}
}
\description{
List Packages Installed Inside Environment
}
