% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_packages}
\alias{list_packages}
\title{List Packages Installed Inside Environment}
\usage{
list_packages(env_name = "condathis-env")
}
\arguments{
\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to 'condathis-env'. If the specified environment does not exist, it will be created automatically using create_env() function from the \code{condathis} package.}
}
\description{
List Packages Installed Inside Environment
}
