% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List Packages Installed Inside Environment}
\usage{
list_packages(env_name = "condathis-env", verbose = FALSE)
}
\arguments{
\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to 'condathis-env'. If the specified environment does not exist, it will be created automatically using create_env() function from the \code{condathis} package.}

\item{verbose}{Logical. Should command and messages be print to
the console.
Defaults to \code{TRUE}.}
}
\description{
List Packages Installed Inside Environment
}
