% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List Packages Installed Inside Environment}
\usage{
list_packages(env_name = "condathis-env", verbose = FALSE)
}
\arguments{
\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to \code{"condathis-env"}.
If the specified environment does not exist, it will be created automatically using \code{create_env()}.}

\item{verbose}{Logical. Whether to display detailed output messages. Defaults to \code{FALSE}.}
}
\description{
List Packages Installed Inside Environment
}
