% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micromamba_bin_path.R
\name{micromamba_bin_path}
\alias{micromamba_bin_path}
\title{Retrieve the path to the \code{condathis} managed installation of \code{micromamba}.}
\usage{
micromamba_bin_path()
}
\description{
Retrieve the path to the \code{condathis} managed installation of \code{micromamba}.
}
