% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_cmd.R
\name{native_cmd}
\alias{native_cmd}
\title{Run Micromamba Command}
\usage{
native_cmd(
  conda_cmd,
  conda_args = NULL,
  ...,
  verbose = FALSE,
  error = c("cancel", "continue"),
  stdout = "|",
  stderr = "|"
)
}
\arguments{
\item{conda_cmd}{Character. Conda subcommand to be run.
E.g. "create", "install", "env", "--help", "--version".}

\item{conda_args}{Character vector. Additional arguments passed to
the Conda command.}

\item{...}{Additional arguments to be passed to the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special characters or spaces.}

\item{verbose}{Logical. Whether to display detailed output messages. Defaults to \code{FALSE}.}

\item{error}{Character string. How to handle errors. Options are \code{"cancel"} or \code{"continue"}. Defaults to \code{"cancel"}.}

\item{stdout}{Default: "|" keep stderr to the R object
returned by \code{run()}.
A character string can be used to define a file path to be used as standard error. e.g: "error.txt".}
}
\description{
Run a command using micromamba executable in the native backend.
}
