% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{packages_search_docker}
\alias{packages_search_docker}
\title{Search for Packages in Channels using Docker}
\usage{
packages_search_docker(
  packages,
  channels = c("bioconda", "conda-forge"),
  additional_channels = NULL,
  container_name = "condathis-micromamba-base",
  image_name = "luciorq/condathis-micromamba:latest"
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge")' are used for solving
dependencies.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}

\item{container_name}{Character. Name of the Container created by Docker.
Defaults to \code{"condathis-micromamba-base"}.}

\item{image_name}{Character. Name of the Docker Image used, it will try
to pull it automatically if internet connection is available.
Defaults to \code{"luciorq/condathis-micromamba:latest"}.}
}
\description{
Search for Packages in Channels using Docker
}
