% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{packages_search_native}
\alias{packages_search_native}
\title{Search For Package in Channels}
\usage{
packages_search_native(
  packages,
  channels = c("bioconda", "conda-forge", "defaults"),
  method = c("native", "docker", "singularity"),
  additional_channels = NULL
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge", "defaults")' are used for solving
dependencies.}

\item{method}{Character. Beckend method to run \code{micromamba}, the default is
"auto" running "native" with the \code{micromamba} binaries installed
by \code{condathis}. Additional methods are supported for using Docker Linux Containers
"docker" and Singularity Containers "singularity" as the beckends,
those are especially useful if running on systems where the Conda
recipes are not available for the OS or CPU architecture in place.
The container-based backends leverage \code{dockerthis} R package.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}
}
\description{
Search if Package is available in required Channels
}
