% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{packages_search_native}
\alias{packages_search_native}
\title{Search For Package in Channels}
\usage{
packages_search_native(
  packages,
  channels = c("bioconda", "conda-forge"),
  platform = NULL,
  additional_channels = NULL,
  verbose = FALSE
)
}
\arguments{
\item{packages}{Character vector. Names of the packages, and
version strings if necessary, e.g. 'python=3.11'. The use of the \code{packages}
argument assumes that env_file is not used.}

\item{channels}{Character vector. Names of the channels to be included.
By default 'c("bioconda", "conda-forge")' are used for solving
dependencies.}

\item{platform}{Character. Platform to search for \code{packages}.
Defaults to \code{NULL} which will use the current platform.
E.g. "linux-64", "linux-32", "osx-64", "win-64", "win-32", "noarch".
Note: on Apple Silicon MacOS will use "osx-64" instead of "osx-arm64"
if Rosetta 2 is available.}

\item{additional_channels}{Character. Additional Channels to be added to the
default ones.}

\item{verbose}{Logical. Should command and messages be print to
the console.
Defaults to \code{TRUE}.}
}
\description{
Search if Package is available in required Channels
}
