% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_method.R
\name{packages_search_singularity}
\alias{packages_search_singularity}
\title{Search for packages in Channels using Singularity}
\usage{
packages_search_singularity(
  packages,
  channels = c("bioconda", "conda-forge", "defaults"),
  additional_channels = NULL,
  sif_image_path = NULL
)
}
\description{
Search for packages in Channels using Singularity
}
