% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{parse_output}
\alias{parse_output}
\title{Parse the output of \code{\link[=run]{run()}}}
\usage{
parse_output(res, stream = c("stdout", "stderr"))
}
\arguments{
\item{res}{Result from \code{\link[=run]{run()}}.}

\item{stream}{Data stream to parse.}
}
\description{
Parse the output of \code{\link[=run]{run()}}
}
