% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_env.R
\name{remove_env}
\alias{remove_env}
\title{Remove a Conda environment}
\usage{
remove_env(
  env_name = "condathis-env",
  verbose = c("silent", "cmd", "output", "spinner", "full")
)
}
\arguments{
\item{env_name}{Character string with the environment name to remove.
Defaults to \code{"condathis-env"}.}

\item{verbose}{Character string controlling console output.
Supported values are \code{"silent"}, \code{"cmd"}, \code{"output"}, \code{"spinner"},
and \code{"full"}. Defaults to \code{"silent"}.}
}
\value{
A process result list (from \code{processx::run()}) with command output,
error output, exit status, and timeout information.
}
\description{
Removes an environment managed by \code{condathis}.
}
\examples{
\dontrun{
condathis::with_sandbox_dir({
  condathis::create_env(
    packages = "bioconda::fastqc",
    env_name = "fastqc-env"
  )
  condathis::remove_env(env_name = "fastqc-env")
})
}

}
