% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_env.R
\name{remove_env}
\alias{remove_env}
\title{Remove a Conda Environment}
\usage{
remove_env(env_name = "condathis-env", verbose = TRUE)
}
\arguments{
\item{env_name}{Character. Name of the Conda environment where the packages
are going to be installed. Defaults to 'condathis-env'.}

\item{verbose}{Logical. Should command and messages be print to
the console.
Defaults to \code{TRUE}.}
}
\description{
Remove a Conda environment previously created by \code{create_env()}.
}
