% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run Command Line tools in a Conda environment.}
\usage{
run(cmd, ..., env_name = "condathis-env")
}
\arguments{
\item{cmd}{Character. The main command to be executed in the Conda environment.}

\item{...}{Additional arguments to be used in the command. These arguments will be passed directly to the command executed in the Conda environment.
File paths should not contain special character or spaces.}

\item{env_name}{Character. The name of the Conda environment where the tool will be run. Defaults to 'condathis-env'. If the specified environment does not exist, it will be created automatically using create_env() function from the \code{condathis} package.}
}
\description{
This function allows the execution of command line tools within a specific Conda environment.
It runs the provided cmd command in the designated Conda environment,
using the Micromamba binaries managed by the \code{condathis} package.
}
\examples{
## Run a simple command in the default Conda environment
run("ls", "-l")

## Run a command in a specific Conda environment
run("python script.py", env_name = "my-conda-env")

## Run a command with additional arguments
run("my-command", "--arg1", "--arg2=value", env_name = "my-conda-env")
}
\seealso{
\code{\link{install_micromamba}}, \code{\link{create_env}}
}
