% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bin.R
\name{run_bin}
\alias{run_bin}
\title{Run a binary without environment activation}
\usage{
run_bin(
  cmd,
  ...,
  env_name = "condathis-env",
  verbose = c("output", "silent", "cmd", "spinner", "full"),
  error = c("cancel", "continue"),
  stdout = "|",
  stderr = "|",
  stdin = NULL
)
}
\arguments{
\item{cmd}{Character string with the command to execute.}

\item{...}{Additional unnamed command arguments passed to \code{cmd}.}

\item{env_name}{Character string with the target environment name.
Defaults to \code{"condathis-env"}.}

\item{verbose}{Character string controlling console output.
Supported values are \code{"output"}, \code{"silent"}, \code{"cmd"}, \code{"spinner"},
and \code{"full"}. Defaults to \code{"output"}.}

\item{error}{Character string that controls error behavior.
Supported values are \code{"cancel"} and \code{"continue"}.
Defaults to \code{"cancel"}.}

\item{stdout}{Standard output target.
Defaults to \code{"|"} (capture stdout in the returned object).
Provide a file path to redirect stdout to a file.}

\item{stderr}{Standard error target.
Defaults to \code{"|"} (capture stderr in the returned object).
Provide a file path to redirect stderr to a file.}

\item{stdin}{Standard input source.
Defaults to \code{NULL} (no stdin stream).
Provide a file path to use file contents as stdin.}
}
\value{
A process result list (from \code{processx::run()}) with command output,
error output, exit status, and timeout information.
}
\description{
Executes a binary using files from a target Conda environment, but without
running environment activation scripts.
This is a lower-level execution mode than \code{run()}.
}
\examples{
\dontrun{
condathis::with_sandbox_dir({
  # Example assumes that 'my-env' exists and contains 'python'
  # Run 'python' with a script in 'my-env' environment
  condathis::run_bin(
    "python", "-c", "import sys; print(sys.version)",
    env_name = "my-env"
  )

  # Run 'ls' command with additional arguments
  condathis::run_bin("ls", "-la", env_name = "my-env")
})
}

}
