% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symlink_micromamba_bin.R
\name{symlink_micromamba_bin}
\alias{symlink_micromamba_bin}
\title{Create a Symlink to the \code{micromamba} Executable}
\usage{
symlink_micromamba_bin(path = NULL, force = FALSE)
}
\arguments{
\item{path}{A character string specifying the path to the \code{micromamba} binary
to symlink. If \code{NULL}, the function attempts to locate a user-installed binary.}

\item{force}{A logical value indicating whether to overwrite an existing symlink.
Defaults to \code{FALSE}.}
}
\value{
Invisibly returns the path to the \code{micromamba} symlink created or verified.
}
\description{
This function checks whether the \code{micromamba} binary is already available on the
system's \code{PATH}. If not, it creates a symbolic link to the binary managed by \code{condathis}.
Users can specify a custom path to an existing \code{micromamba} binary or force
the creation of a new symlink.
}
\keyword{internal}
